requirements.txt
python-nmap3
netifaces